removing jobConfs external constructor after Markus comments
